Move common EnglishAuction types to a shared module
Problem: Some types, like `AuctionId`, are used in all versions of the
English Auction contracts, and they're either duplicated or re-exported
multiple times.

Solution: move them to a shared module.
