+++ packages/minter-contracts/src-hs/Lorentz/Contracts/EnglishAuction/FA2.hs
@@ -6,6 +6,7 @@ import Lorentz
 import Fmt (Buildable(..), genericF)
 import qualified Lorentz.Contracts.AllowlistSimple as AllowlistSimple
 import qualified Lorentz.Contracts.AllowlistToken as AllowlistToken
+import Lorentz.Contracts.EnglishAuction.Common
 import Lorentz.Contracts.MinterSdk
 import qualified Lorentz.Contracts.NoAllowlist as NoAllowlist
 import Lorentz.Contracts.PausableAdminOption
@@ -15,31 +16,6 @@ import qualified Michelson.Typed as T
 
 -- Types
 ----------------------------------------------------------------------------
-
-newtype AuctionId = AuctionId Natural
-  deriving stock (Generic, Eq, Ord, Show)
-  deriving newtype (IsoValue, HasAnnotation, Buildable)
-
-data FA2Token = FA2Token
-  { tokenId :: TokenId
-  , amount :: Natural
-  }
-  deriving stock (Eq, Show)
-
-customGeneric "FA2Token" ligoCombLayout
-deriving anyclass instance IsoValue FA2Token
-deriving anyclass instance HasAnnotation FA2Token
-instance Buildable FA2Token where build = genericF
-
-data Tokens = Tokens
-  { fa2Address :: Address
-  , fa2Batch :: [FA2Token]
-  }
-
-customGeneric "Tokens" ligoCombLayout
-deriving anyclass instance IsoValue Tokens
-deriving anyclass instance HasAnnotation Tokens
-instance Buildable Tokens where build = genericF
 
 data BidCurrency = BidCurrency
   { fa2Address :: Address
