+++ packages/minter-contracts/test-hs/Test/EnglishAuction/AllowlistedTokenFA2.hs
@@ -11,6 +11,7 @@ import Morley.Nettest
 import Tezos.Address
 
 import Lorentz.Contracts.EnglishAuction.Allowlisted
+import Lorentz.Contracts.EnglishAuction.Common
 import Lorentz.Contracts.EnglishAuction.FA2
 import Test.EnglishAuction.Util
 import Test.Util
