+++ packages/minter-contracts/test-hs/Test/EnglishAuction/AllowlistedTokenTez.hs
@@ -9,6 +9,7 @@ import Lorentz.Value
 import Morley.Nettest
 
 import Lorentz.Contracts.EnglishAuction.Allowlisted
+import Lorentz.Contracts.EnglishAuction.Common
 import Lorentz.Contracts.EnglishAuction.Tez
 import Test.EnglishAuction.Util
 import Test.Util
