+++ packages/minter-contracts/test-hs/Test/EnglishAuction/FA2FixedFee.hs
@@ -10,8 +10,10 @@ import qualified Hedgehog.Gen as Gen
 import qualified Hedgehog.Range as Range
 import qualified Indigo.Contracts.FA2Sample as FA2
 import Lorentz hiding (amount, balance, contract, now)
+import Lorentz.Contracts.EnglishAuction.Common
 import Lorentz.Contracts.EnglishAuction.FA2FixedFee
 import Lorentz.Contracts.MinterSdk
+import qualified Lorentz.Contracts.NoAllowlist as NoAllowlist
 import Lorentz.Contracts.Spec.FA2Interface (TokenId(..))
 import qualified Lorentz.Contracts.Spec.FA2Interface as FA2
 import Michelson.Typed (convertContract, untypeValue)
@@ -571,7 +573,7 @@ genBid testData@TestData{testMinRaise} previousBid = do
 ----------------------------------------------------------------------------
 
 data Setup = Setup
-  { contract :: TAddress AuctionEntrypoints
+  { contract :: TAddress (AuctionEntrypoints NoAllowlist.Entrypoints)
   , fa2Contracts :: [TAddress FA2.FA2SampleParameter]
   , bidFA2Contract :: TAddress FA2.FA2SampleParameter
   , bidFA2ContractAdmin :: Address
@@ -609,6 +611,7 @@ testSetup testData = do
             , tokenId = testBidTokenId testData
             }
       , auctions = mempty
+      , allowlist = ()
       , fee = FeeData feeCollector (testFeePercent testData)
       }
 
@@ -639,9 +642,9 @@ waitForAuctionToStart :: (HasCallStack, MonadNettest caps base m) => TestData ->
 waitForAuctionToStart TestData{testTimeToStart} =
   advanceTime (sec $ fromIntegral testTimeToStart)
 
-originateAuctionContract :: MonadNettest caps base m => AuctionStorage -> m (TAddress AuctionEntrypoints)
+originateAuctionContract :: MonadNettest caps base m => AuctionStorage -> m (TAddress (AuctionEntrypoints NoAllowlist.Entrypoints))
 originateAuctionContract storage = do
-  TAddress @AuctionEntrypoints <$> originateUntypedSimple "auction-tez-fixed-fee"
+  TAddress @(AuctionEntrypoints NoAllowlist.Entrypoints) <$> originateUntypedSimple "auction-tez-fixed-fee"
     (untypeValue $ toVal storage)
     (convertContract englishAuctionFA2FixedFeeContract)
 
@@ -723,14 +726,14 @@ configureAuction testData Setup{fa2Contracts, contract, startTime, endTime}  = d
     , endTime = endTime
     }
 
-placeBid :: (HasCallStack, MonadNettest caps base m) => TAddress AuctionEntrypoints -> Natural -> m ()
+placeBid :: (HasCallStack, MonadNettest caps base m) => TAddress (AuctionEntrypoints NoAllowlist.Entrypoints) -> Natural -> m ()
 placeBid contract bidAmount =
   call contract (Call @"Bid") (BidParam (AuctionId 0) bidAmount)
 
-cancelAuction :: (HasCallStack, MonadNettest caps base m) => TAddress AuctionEntrypoints -> m ()
+cancelAuction :: (HasCallStack, MonadNettest caps base m) => TAddress (AuctionEntrypoints NoAllowlist.Entrypoints) -> m ()
 cancelAuction contract =
   call contract (Call @"Cancel") (AuctionId 0)
 
-resolveAuction :: (HasCallStack, MonadNettest caps base m) => TAddress AuctionEntrypoints -> m ()
+resolveAuction :: (HasCallStack, MonadNettest caps base m) => TAddress (AuctionEntrypoints NoAllowlist.Entrypoints) -> m ()
 resolveAuction contract =
   call contract (Call @"Resolve") (AuctionId 0)
